refactor(server): split routes-api.ts monolith into routes-api/ domain modules
The 1093-line routes-api.ts violated the repo's own ≤1000-line invariant
(ARCHITECTURE.md §IV — which also still described it as "~670 lines").

Split the single handleApi if-chain into a thin index.ts orchestrator + per-domain
handlers (auth/servers/agents/channels/messages/tasks/reminders/attachments), with
ctx.ts (typed per-gate context) and shared.ts (attachMentions/userChannels). Route
bodies are moved VERBATIM; only the gate context is threaded via a ctx object instead
of closure vars. Pure refactor — zero behavior change, zero new/removed endpoints.

The orchestrator preserves the three auth gates (public → +userId → +serverId) and the
gate-2 dispatch order, both security-load-bearing (a route behind the wrong gate = an
authz bug; first-match order disambiguates e.g. /channels/saved before /channels/:id).

Verified (same seeded DB, pre- vs post-split):
- npm run typecheck (root + web) green
- route-inventory diff: 78 route guards, set-identical (none lost / added / method-changed)
- behavioral curl regression: 48 calls across every domain + both gate boundaries +
  collision probes — byte-for-byte identical responses
- moved routes (runtime-models, avatars, reconnect) reachable

Docs: ARCHITECTURE.md codemap + ≤1000-line invariant updated.
