+++ ARCHITECTURE.md
@@ -37,7 +37,7 @@ Agent data plane   Agent process → HTTP /agent-api/*  (Bearer per-agent token 
 ### Control plane + data plane (`src/server/`, TypeScript — the production implementation)
 
 - `index.ts` — HTTP entry point: route order `/agent-api` → `/api` → static `web/dist` (SPA fallback); mounts socket.io (human real-time), bare WebSocket (daemon control plane), reminder scheduler.
-- `routes-api.ts` — **Human-side REST `/api/*`** (~670 lines, largest file): auth (dev-login / register / login + `invite-info` / `accept-invite` via join-link), servers (`POST /servers` creates workspace [core.createServer: server + owner + default #all] + `POST /servers/:id/avatar` + `GET /servers/:sid/members/:uid/profile` + join-links CRUD [manageMembers] + `PATCH/DELETE /servers/:id/members/:uid` role management/removal [changeMemberRoles/manageMembers; constraints: not self + preserve last owner]; `GET /servers` returns role + capabilities per item) / members (with description) / machines (incl. `POST …/machines/:id/reconnect` [manageMachines] — rotate the connection key on an existing **offline** row so a machine whose one-time key was lost can rejoin without a duplicate; refuses an online machine) / notification-settings / sidebar-order, agents (start/stop/reset/workspace-files/activity-log/scopes/skills), channels (inbox/unread/saved/threads/dm/files), `GET /mentions` (message-grained @-of-me activity stream, read or unread; `limit`/`offset` + `hasMore` load-more paging), messages (`GET messages/channel/:id` keyset-paginated: `before=<seq>` cursor → older page + `hasMore`)/sync/search, tasks, `actions/:id/mark-executed` (action card human-in-the-loop submit), attachments, reminders, announcements.
+- `routes-api/` — **Human-side REST `/api/*`** (directory: a thin `index.ts` orchestrator runs the three auth gates [public → +userId → +serverId], builds a per-gate `ctx` [`ctx.ts`], and dispatches to per-domain handlers — `auth`/`servers`/`agents`/`channels`/`messages`/`tasks`/`reminders`/`attachments`, shared helpers in `shared.ts`; **gate order + gate-2 dispatch order are security-load-bearing** — see header comment in `index.ts`). Endpoints by group: auth (dev-login / register / login + `invite-info` / `accept-invite` via join-link), servers (`POST /servers` creates workspace [core.createServer: server + owner + default #all] + `POST /servers/:id/avatar` + `GET /servers/:sid/members/:uid/profile` + join-links CRUD [manageMembers] + `PATCH/DELETE /servers/:id/members/:uid` role management/removal [changeMemberRoles/manageMembers; constraints: not self + preserve last owner]; `GET /servers` returns role + capabilities per item) / members (with description) / machines (incl. `POST …/machines/:id/reconnect` [manageMachines] — rotate the connection key on an existing **offline** row so a machine whose one-time key was lost can rejoin without a duplicate; refuses an online machine) / notification-settings / sidebar-order, agents (start/stop/reset/workspace-files/activity-log/scopes/skills), channels (inbox/unread/saved/threads/dm/files), `GET /mentions` (message-grained @-of-me activity stream, read or unread; `limit`/`offset` + `hasMore` load-more paging), messages (`GET messages/channel/:id` keyset-paginated: `before=<seq>` cursor → older page + `hasMore`)/sync/search, tasks, `actions/:id/mark-executed` (action card human-in-the-loop submit), attachments, reminders, announcements.
 - `routes-agent.ts` — **Agent data plane `/agent-api/*`**: full agent operation surface — message (check/send/react/read/resolve), server/info (channels + agents + humans, members include `description` field so agents understand teammates), channel (join/members/leave), task (list/claim/update/new/unclaim), thread (reply/read/unfollow), search, attachment (upload/view), profile (show/update), reminder (schedule/list/cancel/snooze), `action/prepare` (send B-mode quick-commit card, scope `action:prepare`; variants: channel:create / agent:create).
 - `core.ts` — **Message core**: seq allocation, `@mention` parsing (stored structurally in `message_mentions` table), persistence, human-side broadcast, agent wake delivery, target resolution. All "send a message" business logic converges here; S→D commands `agent:start/stop/deliver/sleep/...` also originate here via daemonHub.
 - `auth.ts` — **Two auth stacks**: human JWT (`verifyUser`, scrypt password storage) + agent (`resolveAgent`: Bearer = per-agent token, SHA-256 compared against `agents.agentTokenHash` + header `x-agent-id`; token must belong to that agentId → prevents cross-agent / cross-server impersonation; no master key). ⚠️ Daemon WS `?key=` handshake auth (machine key) is **inline in `ws.ts`**, completely separate from agent credentials.
@@ -138,7 +138,7 @@ cd web && npm run dev  # Vite HMR dev server
 - **`storageKey` is driver-agnostic**: switching local ↔ S3 requires no changes to business logic.
 - **Agent communication is unified through the `open-tag` CLI** (generated at `~/.open-tag/bin`, injected into the agent's PATH). Claude agents do **not** use MCP. What the agent cannot see in its context effectively does not exist — knowledge is persisted to the workspace `MEMORY.md`.
 - **Agents are single-threaded / one agent = one session (intentional design, not a limitation)**: one agent = one process, one session; turns are **serial** (while busy, deliver-debounce batches incoming messages rather than interrupting). **Parallelism lives at the team level (multiple agents collaborating), not inside a single agent** — "employee mental model" (see §I): scale by hiring more people, not by making one person multi-thread. `Task` and sub-agents are **not blocked** (inherited from the host machine's tooling), but the standing prompt does not encourage their use → **do not build logic that depends on which sub-agents happen to be installed, and do not add `/fork`-style intra-agent parallelism**.
-- **Single file ≤ 1000 lines** (split if exceeded). Currently the largest is `routes-api.ts` at ~670 lines.
+- **Single file ≤ 1000 lines** (split if exceeded). The former 1093-line `routes-api.ts` is now the `routes-api/` directory (largest module `channels.ts` ~310); the largest single file is now `core.ts` (~630).
 
 ## IV. Boundaries (made explicit because they are easy to miss)
