+++ src/server/index.ts
@@ -4,7 +4,7 @@ import http from "node:http";
 import { readFile } from "node:fs/promises";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
-import { handleApi } from "./routes-api.js";
+import { handleApi } from "./routes-api/index.js";
 import { handleAgentApi } from "./routes-agent.js";
 import { attachWs } from "./ws.js";
 import { attachSocketIO } from "./socketio.js";
